backend/requirements.txt
aiohttp==3.8.3
allure_python_commons==2.12.0
backend==0.2.4.1
fastapi==0.92.0
filetype==1.2.0
Jinja2==3.1.2
jmespath==0.9.5
jose==1.0.0
loguru==0.4.1
paramiko==2.11.0
passlib==1.7.4
pydantic==1.8.2
PyJWT==2.6.0
pytest==7.2.0
PyYAML==6.0
requests==2.28.1
requests_toolbelt==0.10.1
sentry_sdk==0.14.4
setuptools==49.2.1
SQLAlchemy==1.3.22
starlette==0.16.0
tortoise==0.1.1
urllib3==1.26.2
uvicorn==0.15.0
